pip_requirements.txt
anaconda-client==1.7.2
appdirs==1.4.3
appnope==0.1.0
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
backcall==0.1.0
bleach==2.1.4
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
clyent==1.2.2
constantly==15.1.0
cryptography==2.3.1
cycler==0.10.0
decorator==4.3.0
entrypoints==0.2.3
html5lib==1.0.1
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.1
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.3
mistune==0.8.3
mkl-fft==1.0.4
mkl-random==1.0.1
nb-anacondacloud==1.4.0
nb-conda==2.2.1
nb-conda-kernels==2.1.0
nbconvert==5.3.1
nbformat==4.4.0
nbpresent==3.0.2
notebook==5.6.0
numpy==1.15.1
olefile==0.46
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
prometheus-client==0.3.1
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
Pygments==2.2.0
pyOpenSSL==18.0.0
pyparsing==2.2.0
PySocks==1.6.8
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
pyzmq==17.1.2
requests==2.19.1
scikit-learn==0.19.2
scipy==1.1.0
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
torch==0.4.1
torchvision==0.2.1
tornado==5.1
traitlets==4.3.2
Twisted==17.5.0
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.1
zope.interface==4.5.0
